// src/modules/notifications/enums/notification-priority.enum.ts
export enum NotificationPriority {
  LOW = 'LOW',
  NORMAL = 'NORMAL',
  HIGH = 'HIGH',
}

export enum EmailTemplateType {
  ACCOUNT_VERIFICATION = 'ACCOUNT_VERIFICATION',
  ADMIN_NEW_USER_NOTIFICATION = 'ADMIN_NEW_USER_NOTIFICATION',
  ADMIN_NEW_ORGANIZATION_NOTIFICATION = 'ADMIN_NEW_ORGANIZATION_NOTIFICATION',
  DOCUMENT_SUBMISSION_NOTIFICATION = 'DOCUMENT_SUBMISSION_NOTIFICATION',
  DOCUMENT_STATUS_UPDATE = 'DOCUMENT_STATUS_UPDATE',
  PASSWORD_RESET = 'PASSWORD_RESET',
}